move test toggle menu bar
